Preproc mods- added try-catch for figs and commnts
Changed preproc fig creation to 'try' creating each figure and if it fails
to just copy a blank figure to the output directory with the same name as
the figure that would have been created. The purpose of this is to keep 6
figures per station, and have the LaTeX output 6 figures per page, which
results in one page per station.

Adjusted comments so output to screen was pretty