Fixed bug: series in plot did not match legend
Discovered a bug where the series in the plots did not match the series in
the legend. The problem was that the STATION map used the key order for
alternatives (which was sorted alphabetically) and the legends used the
order in the INI.MODEL_RUN_DESC variable, which was the order the user
stated.
The issue did not appear as long as the user's Alt order was alphabetical,
(ie M01, M06), but appeared when it was not alphabetical (ie M06, M01).