updated transect read scripts
- Removed bad hardcode where load_XL_GRID would use only the filename for
the 3DSZQ file, not the OL file. (It has worked thus far because we use
the same file for both).

- Removed functionality to read transect definitions from MATLAB file
because it doesn't actually save any time to do so and complicates the
script.

- Removed functionality to save transect definitions to file. It doesn't
work correctly because we are calling this script twice (once for OL and
once for 3DSZQ) so the file would get overwritten incorrectly.

- Adjusted variable names to more descriptive names.