Update to use nearest chainage if exact not found
commit author: Georgio Tachiev

commit message:

Updated to use the nearest chainages:

This update resolves:

.Uses the nearest chainages to avoid discrepancies between using MIKE 11
and MIKE 1D which appeared in MIKE 2019 software. The discharges are
reported differently for MIKE 1D and MIKE 11 which requires code to be
able to read both versions without errors

.The loop over all stations have been updated to resolve the reading
issues and to improve efficiency of reading all files.

NOTE: This was originally a branch called REVISE_GENERATE_FOR_CHAINAGES
and all edits were written by Georgio. This commit just copies the edits
to a more recent commit branch to keep them fresh at the top of the edit
chain.