Transects - update to sum over ALL layers
Updated scripts to sum transect data over ALL layers by default, instead
of having to specify each layer you want data for in the Excel file.

The Excel file defining the transects has been updated, and the column
specifying the layer has been made inactive.

The backstory:
When the model was only 3 layers, it seemed cool to have the flexibility
to specify which layers we wanted to sum transect data for. The fact
that we had to list each cell 3 times in the Excel file (one for each
layer) was a small price to pay for this flexibility. As the model grew
to 5 layers, the unmanagibility of listing cells 5 times over became too
much, so this extra functionality was abandoned in favor of having more
managable code. Also, as we start playing with the number of layers in
the model, this makes it MUCH simpler to compare model output.