Moved Sample_Input/Output to ENP_FILES
Changed input and output directory locations on working scripts to point
to new file location in new master directory called ENP_FILES.

Changed scripts so they would not overwrite any files in
ENP_TOOLS_Sample_Input. This should now be read only.

Changed scripts to have two testing options - one is a 'unit' test where
the script will read only from files in ENP_TOOLS_Sample_Input and write
only to it's own subdirectory in ENP_TOOLS_Output. The second is a
'sequential' test where the input and output directories are the same
files, and the output of one script can be used as the input to the next
script to test the scripts work together (I/O directory is calles
'ENP_TOOLS_Output_Sequential').

Although not part of the commit, I added two output directories to
ENP_FILES that have the current expected output of the scripts, that the
user can compare their output against to insure they are working
correctly.

For scripts that are not working at the moment, I added a print message
that the script is not currently working that will appear first on the
output screen.