res11read - added switch for old/new method
Added switch to allow user to select between the new method of reading M11 data form the .res11 file, or the old method of reading M11 data from the dfs0 file. Currently the new method is much slower than the old method, but more accurate.

Updated readFileCompCoord.m that reads M11 chainage values from the Excel file, to either read from the original column (for the old method) or from a new column of chainages (for the new method).

While updating readFileCompCoord.m, an error was found where data was being read from either the RAW dataset or the TXT dataset. The problem was that these datasets had different numbers of rows, so the index 'i' used to reference the RAW dataset would have needed to be offset to read the same row in the TXT dataset.

The MATLAB help file also said that using xlsread was not advised starting in R2019a. This was not addressed.