Add edits from master to facilitate merge
This branch and the master branch diverged quite a bit while being
developed, so to facilitate resolution of impending merge conflicts
several edits from the master branch were copied into this branch.

This will allow us to 'catch up' with the edits in the master branch,
and begin a new branch with upcoming edits on this res11 read feature.

Also, because the new res11 read feature is not quite complete, it was
commented out of the active code with the comment %BETA%.

Spacing and indentation were also modified.