Fix an error causing inconsistent *.png size
When maps are generated in the following scripts and they saved the map png files, they would not be saved at the correct resolution due to properties in the figure from previous scripts. Using the close command will let a new figure window be opened with all default properties, and the final png will be saved at the correct resolution.