adjust header info in flagged *.dat files
In the case that *.dat file has its values converted from the NAVD88 datum to the NGVD29, the header will now have the datum as NGVD29 instead of NAVD88 with a note that the values were converted. The ground surface elevation in the header will also be converted from NAVD88 to NGVD29.