fix an issue with null values being replaced
This change involves the case where flagged *.dat files are written if a value was null in the raw *.dat file and triggers the "not a number" flag. When the value is written to the output flagged *.dat file a null value will be written. When written to the dfs0 file, the value will be replaced with the no data value of -1e-35.