fix 0 file size error and error handling
Added an additional check where the dfs0 and res11 files will only be read if they exist and are non zero in file size. Additionally found an error where if a file read error occurred, the INI variable was replaced with the Error object and would cause subsequent scripts to fail.