preproc - changed BC2D extrapolation method
.. and other small changes

BC2D scatteredinterpolant did not specify extrapolation method.
I added this so it wsa explicit, and changed it from 'linear' to a
nearest neighbor technique.

This happened in conjunction with the creation of 'fake' boundary
conditions on the fringes of the model grid where there was no input
data, and the interpolation would get wacky.