preproc - overhauled DFE data handling of datum
Major overhaul of preproc_flag_DFE_file script.

Previous edits didn't quite fix the issues we were having.

Currently the DFE database is in a state of flux between NGVD29 and
NAVD88, calusing all sorts of issues with correctly processing the data,
and knowing what datum each dataset and station elevation is in.
They do, for the most part, follow a complicated set of rules, that
needed to be coded into this script.

Once the DFE database is fully transitioned, most of this code can and
probably should be stripped out.

See preproc_flag_DFE_file.m and email chains around 12/2020 for more
details if needed.