preproc - add BC2D non-interpolated option
added option to BC2D scripts to not interpolate timeseries.
Before we had two interpolation options for temporal interpolation of
timeseries at each station where there was missing data. The Fourier
fits a curve and fills the station's missing data based on the fit. The
Julian method filles the timeseries based on Julian day averages. This
commit adds the option to _not_ fill the timeseries with interpolated
data.
The script then produces a spatially interpolated surface based on the
filled (or unfilled) timeseries.